refactor(app): getHighestPriorityError to xforms
