+++ app/src/transformations/runs/index.ts
@@ -3,3 +3,4 @@ export * from './getRunTimeParameterFilesForRun'
 export * from './getLatestCurrentOffsets'
 export * from './formatTimestamp'
 export * from './onDeviceDisplayFormatTimestamp'
+export * from './getHighestPriorityError'
